Valorisation persistée et immuable, avec trace complète (KAI-203)
Assemble les trois moteurs : ajustement de set, pondération, détection
d'anomalies, cote pondérée et indice de confiance. Chaque appel crée une
version ; une valorisation n'est jamais écrasée (règle 4), ce qu'un test
vérifie en constatant le refus du trigger d'append-only.

Chaque comparable est figé dans `valuation_comparables` avec
l'intégralité de ses facteurs — y compris ceux exclus par décision
humaine ou signalés comme anomalie, avec leur motif. La cote reste ainsi
rejouable même si les comparables évoluent ensuite.

Les comparables exclus par un humain ne participent pas non plus à la
détection d'anomalies : une exclusion motivée ne doit pas déplacer
indirectement le seuil qui écarte les autres.

Le barème figé est stocké depuis son texte JSON source. Le sérialiseur
refusait les Decimal, et les convertir en flottants aurait détruit la
précision de l'instantané censé rendre la cote rejouable.

Corrige par ailleurs 76 colonnes Numeric annotées `float` dans les
modèles : SQLAlchemy renvoie des Decimal, et l'annotation contredisait
la règle 2 tout en masquant les erreurs de type réelles.

Deux lectures que la spécification ne tranche pas sont signalées en
Q-12 : la dimension d'état gouvernant la similarité — l'écart le plus
défavorable est retenu — et l'écart de set supérieur à un niveau, qui
retombe sur le coefficient le plus bas.

+++ apps/api/app/shared/infrastructure/db/models/analyses.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import ForeignKey, Integer, Numeric, Text, text
 from sqlalchemy.dialects.postgresql import JSONB, TIMESTAMP, UUID
@@ -53,16 +54,16 @@ class Analysis(Base):
     )
     published_at: Mapped[datetime | None] = mapped_column(TIMESTAMP(timezone=True))
 
-    current_price_eur: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    total_cost_eur: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    expected_sale_price_eur: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    raw_max_purchase_price_eur: Mapped[float | None] = mapped_column(Numeric(16, 8))
-    max_purchase_price_eur: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    expected_profit_eur: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    expected_roi: Mapped[float | None] = mapped_column(Numeric(18, 10))
+    current_price_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    total_cost_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    expected_sale_price_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    raw_max_purchase_price_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 8))
+    max_purchase_price_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    expected_profit_eur: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    expected_roi: Mapped[Decimal | None] = mapped_column(Numeric(18, 10))
     expected_days_to_sell: Mapped[int | None] = mapped_column(Integer)
-    score: Mapped[float | None] = mapped_column(Numeric(7, 4))
-    evidence_quality_score: Mapped[float | None] = mapped_column(Numeric(7, 4))
+    score: Mapped[Decimal | None] = mapped_column(Numeric(7, 4))
+    evidence_quality_score: Mapped[Decimal | None] = mapped_column(Numeric(7, 4))
     recommendation: Mapped[str] = mapped_column(
         pg_enum(Recommendation, "recommendation"), nullable=False
     )
