+++ apps/api/app/shared/infrastructure/db/models/platforms.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime, timedelta
+from decimal import Decimal
 
 from sqlalchemy import (
     CHAR,
@@ -60,19 +61,19 @@ class PlatformRule(Base):
     min_poll_interval: Mapped[timedelta | None] = mapped_column(INTERVAL)
     max_poll_interval: Mapped[timedelta | None] = mapped_column(INTERVAL)
 
-    buyer_fee_rate: Mapped[float | None] = mapped_column(Numeric(18, 10))
-    buyer_fee_fixed: Mapped[float | None] = mapped_column(Numeric(16, 2))
+    buyer_fee_rate: Mapped[Decimal | None] = mapped_column(Numeric(18, 10))
+    buyer_fee_fixed: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
     buyer_fee_currency: Mapped[str | None] = mapped_column(CHAR(3))
     buyer_fee_basis: Mapped[str | None] = mapped_column(Text)
-    buyer_fee_min: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    buyer_fee_max: Mapped[float | None] = mapped_column(Numeric(16, 2))
+    buyer_fee_min: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    buyer_fee_max: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
 
-    seller_fee_rate: Mapped[float | None] = mapped_column(Numeric(18, 10))
-    seller_fee_fixed: Mapped[float | None] = mapped_column(Numeric(16, 2))
+    seller_fee_rate: Mapped[Decimal | None] = mapped_column(Numeric(18, 10))
+    seller_fee_fixed: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
     seller_fee_currency: Mapped[str | None] = mapped_column(CHAR(3))
     seller_fee_basis: Mapped[str | None] = mapped_column(Text)
-    seller_fee_min: Mapped[float | None] = mapped_column(Numeric(16, 2))
-    seller_fee_max: Mapped[float | None] = mapped_column(Numeric(16, 2))
+    seller_fee_min: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
+    seller_fee_max: Mapped[Decimal | None] = mapped_column(Numeric(16, 2))
 
     payment_rules: Mapped[dict[str, object]] = mapped_column(
         JSONB, nullable=False, server_default=text("'{}'::jsonb")
